requirements.txt
selenium==4.6.0
requests==2.28.1
webdriver_manager==3.8.4
packaging==21.3
flask-restful==0.3.9
uvicorn==20.1.0